+++ ci/checks/prohibit-bash-usages-from-python.sh
@@ -2,7 +2,7 @@
 
 set -e -o pipefail -u
 
-if git grep -n bash -- tests/ '*.py'; then
+if git grep -n -e bash --and --not -e completion -- tests/ '*.py'; then
   echo
   echo 'Do not rely on bash being installed, not even in tests.'
   echo 'See e.g. https://github.com/VirtusLab/git-machete/pull/929.'