+++ ci/rpm/build-context/run.sh
@@ -10,3 +10,4 @@ fi
 python3 setup.py bdist_rpm
 rpm -i dist/git-machete-*.rpm
 git machete version
+git machete completion bash  # to check if completion files are available in the runtime