+++ ci/snap/ci-deploy.sh
@@ -25,6 +25,7 @@ if [[ ${1-} == "--dry-run" || ${CIRCLE_BRANCH-} != "master" ]]; then
   if command -v git-machete; then exit 1; fi
   sudo snap install git-machete*.snap --dangerous --classic
   git machete version
+  git machete completion bash
   sudo snap remove git-machete
 else
   # Relying on SNAPCRAFT_STORE_CREDENTIALS, provided by the CI